Split implement_dispatch.py (3,469 LOC) into 9 cohesive sibling modules
The god-module is decomposed into: dispatch_helpers (primitives), dispatch_leg (process management), dispatch_bootstrap (step 0), dispatch_ship_seed (step 2 post-dispatch), dispatch_commit_route (checks + commit route + steps 4-6), dispatch_ship (step 8+), dispatch_step18 (step 18 finalize), dispatch_recovery (recovery paths + commit), dispatch_manifest (DispatchState + manifest), and dispatch_step2 (run_dispatch + step2_dispatch). implement_dispatch.py becomes a thin re-export shim preserving the public surface. Updates complexity-baseline.json, ruff.toml, and the two subprocess/env baseline files.
