+++ python/complexity-baseline.json
@@ -2790,97 +2790,97 @@
     "metric": 62
   },
   {
-    "file": "larch/implement/implement_dispatch.py",
-    "code": "C901",
-    "qualified_symbol": "_emit_manifest_invalid_or_recover",
-    "metric": 15
-  },
-  {
-    "file": "larch/implement/implement_dispatch.py",
-    "code": "C901",
-    "qualified_symbol": "commit_main",
-    "metric": 15
-  },
-  {
-    "file": "larch/implement/implement_dispatch.py",
-    "code": "C901",
-    "qualified_symbol": "compute_recovery_paths",
-    "metric": 15
-  },
-  {
-    "file": "larch/implement/implement_dispatch.py",
+    "file": "larch/implement/dispatch_bootstrap.py",
     "code": "C901",
     "qualified_symbol": "step0_bootstrap_main",
     "metric": 14
   },
   {
-    "file": "larch/implement/implement_dispatch.py",
-    "code": "C901",
-    "qualified_symbol": "step2_dispatch_main",
-    "metric": 74
-  },
-  {
-    "file": "larch/implement/implement_dispatch.py",
-    "code": "PLR0911",
-    "qualified_symbol": "_emit_manifest_invalid_or_recover",
-    "metric": 7
-  },
-  {
-    "file": "larch/implement/implement_dispatch.py",
-    "code": "PLR0911",
-    "qualified_symbol": "run_dispatch_main",
-    "metric": 10
-  },
-  {
-    "file": "larch/implement/implement_dispatch.py",
-    "code": "PLR0911",
-    "qualified_symbol": "step2_dispatch_main",
-    "metric": 41
-  },
-  {
-    "file": "larch/implement/implement_dispatch.py",
-    "code": "PLR0912",
-    "qualified_symbol": "_emit_manifest_invalid_or_recover",
-    "metric": 14
-  },
-  {
-    "file": "larch/implement/implement_dispatch.py",
-    "code": "PLR0912",
-    "qualified_symbol": "commit_main",
-    "metric": 15
-  },
-  {
-    "file": "larch/implement/implement_dispatch.py",
+    "file": "larch/implement/dispatch_bootstrap.py",
     "code": "PLR0912",
     "qualified_symbol": "step0_bootstrap_main",
     "metric": 13
   },
   {
-    "file": "larch/implement/implement_dispatch.py",
-    "code": "PLR0912",
-    "qualified_symbol": "step2_dispatch_main",
-    "metric": 79
-  },
-  {
-    "file": "larch/implement/implement_dispatch.py",
-    "code": "PLR0913",
-    "qualified_symbol": "_maybe_mark_step2_telemetry",
-    "metric": 7
-  },
-  {
-    "file": "larch/implement/implement_dispatch.py",
-    "code": "PLR0913",
-    "qualified_symbol": "compute_recovery_paths",
-    "metric": 6
-  },
-  {
-    "file": "larch/implement/implement_dispatch.py",
+    "file": "larch/implement/dispatch_bootstrap.py",
     "code": "PLR0915",
     "qualified_symbol": "step0_bootstrap_main",
     "metric": 78
   },
   {
-    "file": "larch/implement/implement_dispatch.py",
+    "file": "larch/implement/dispatch_helpers.py",
+    "code": "PLR0913",
+    "qualified_symbol": "_maybe_mark_step2_telemetry",
+    "metric": 7
+  },
+  {
+    "file": "larch/implement/dispatch_manifest.py",
+    "code": "C901",
+    "qualified_symbol": "_emit_manifest_invalid_or_recover",
+    "metric": 15
+  },
+  {
+    "file": "larch/implement/dispatch_manifest.py",
+    "code": "PLR0911",
+    "qualified_symbol": "_emit_manifest_invalid_or_recover",
+    "metric": 7
+  },
+  {
+    "file": "larch/implement/dispatch_manifest.py",
+    "code": "PLR0912",
+    "qualified_symbol": "_emit_manifest_invalid_or_recover",
+    "metric": 14
+  },
+  {
+    "file": "larch/implement/dispatch_recovery.py",
+    "code": "C901",
+    "qualified_symbol": "compute_recovery_paths",
+    "metric": 15
+  },
+  {
+    "file": "larch/implement/dispatch_recovery.py",
+    "code": "PLR0913",
+    "qualified_symbol": "compute_recovery_paths",
+    "metric": 6
+  },
+  {
+    "file": "larch/implement/dispatch_recovery.py",
+    "code": "C901",
+    "qualified_symbol": "commit_main",
+    "metric": 15
+  },
+  {
+    "file": "larch/implement/dispatch_recovery.py",
+    "code": "PLR0912",
+    "qualified_symbol": "commit_main",
+    "metric": 15
+  },
+  {
+    "file": "larch/implement/dispatch_step2.py",
+    "code": "PLR0911",
+    "qualified_symbol": "run_dispatch_main",
+    "metric": 10
+  },
+  {
+    "file": "larch/implement/dispatch_step2.py",
+    "code": "C901",
+    "qualified_symbol": "step2_dispatch_main",
+    "metric": 74
+  },
+  {
+    "file": "larch/implement/dispatch_step2.py",
+    "code": "PLR0911",
+    "qualified_symbol": "step2_dispatch_main",
+    "metric": 41
+  },
+  {
+    "file": "larch/implement/dispatch_step2.py",
+    "code": "PLR0912",
+    "qualified_symbol": "step2_dispatch_main",
+    "metric": 79
+  },
+  {
+    "file": "larch/implement/dispatch_step2.py",
     "code": "PLR0915",
     "qualified_symbol": "step2_dispatch_main",
     "metric": 241
