+++ python/env-via-config-constant-baseline.json
@@ -1044,7 +1044,7 @@
     "reason": "grandfathered bare env literal pre-G-Cfg-2 ratchet"
   },
   {
-    "file": "larch/implement/implement_dispatch.py",
+    "file": "larch/implement/dispatch_leg.py",
     "qualified_symbol": "_active_leg_pgid_path",
     "env_name": "IMPLEMENT_TMPDIR",
     "constant": "ENV_IMPLEMENT_TMPDIR",
@@ -1053,7 +1053,7 @@
     "reason": "grandfathered bare env literal pre-G-Cfg-2 ratchet"
   },
   {
-    "file": "larch/implement/implement_dispatch.py",
+    "file": "larch/implement/dispatch_helpers.py",
     "qualified_symbol": "_current_cli_path",
     "env_name": "CLAUDE_PLUGIN_ROOT",
     "constant": "ENV_CLAUDE_PLUGIN_ROOT",
@@ -1062,7 +1062,7 @@
     "reason": "grandfathered bare env literal pre-G-Cfg-2 ratchet"
   },
   {
-    "file": "larch/implement/implement_dispatch.py",
+    "file": "larch/implement/dispatch_helpers.py",
     "qualified_symbol": "_rehydrate_plugin_root",
     "env_name": "CLAUDE_PLUGIN_ROOT",
     "constant": "ENV_CLAUDE_PLUGIN_ROOT",
@@ -1071,7 +1071,7 @@
     "reason": "grandfathered bare env literal pre-G-Cfg-2 ratchet"
   },
   {
-    "file": "larch/implement/implement_dispatch.py",
+    "file": "larch/implement/dispatch_helpers.py",
     "qualified_symbol": "_rehydrate_plugin_root",
     "env_name": "CLAUDE_PLUGIN_ROOT",
     "constant": "ENV_CLAUDE_PLUGIN_ROOT",
@@ -1080,7 +1080,7 @@
     "reason": "grandfathered bare env literal pre-G-Cfg-2 ratchet"
   },
   {
-    "file": "larch/implement/implement_dispatch.py",
+    "file": "larch/implement/dispatch_helpers.py",
     "qualified_symbol": "_tmpdir_from_env",
     "env_name": "IMPLEMENT_TMPDIR",
     "constant": "ENV_IMPLEMENT_TMPDIR",
@@ -1089,7 +1089,7 @@
     "reason": "grandfathered bare env literal pre-G-Cfg-2 ratchet"
   },
   {
-    "file": "larch/implement/implement_dispatch.py",
+    "file": "larch/implement/dispatch_recovery.py",
     "qualified_symbol": "commit_main",
     "env_name": "IMPLEMENT_TMPDIR",
     "constant": "ENV_IMPLEMENT_TMPDIR",
@@ -1098,7 +1098,7 @@
     "reason": "grandfathered bare env literal pre-G-Cfg-2 ratchet"
   },
   {
-    "file": "larch/implement/implement_dispatch.py",
+    "file": "larch/implement/dispatch_recovery.py",
     "qualified_symbol": "commit_main",
     "env_name": "IMPLEMENT_TMPDIR",
     "constant": "ENV_IMPLEMENT_TMPDIR",
@@ -1107,7 +1107,7 @@
     "reason": "grandfathered bare env literal pre-G-Cfg-2 ratchet"
   },
   {
-    "file": "larch/implement/implement_dispatch.py",
+    "file": "larch/implement/dispatch_commit_route.py",
     "qualified_symbol": "commit_route_main",
     "env_name": "IMPLEMENT_TMPDIR",
     "constant": "ENV_IMPLEMENT_TMPDIR",
@@ -1116,7 +1116,7 @@
     "reason": "grandfathered bare env literal pre-G-Cfg-2 ratchet"
   },
   {
-    "file": "larch/implement/implement_dispatch.py",
+    "file": "larch/implement/dispatch_step2.py",
     "qualified_symbol": "run_dispatch_main",
     "env_name": "CLAUDE_PLUGIN_ROOT",
     "constant": "ENV_CLAUDE_PLUGIN_ROOT",
@@ -1125,7 +1125,7 @@
     "reason": "grandfathered bare env literal pre-G-Cfg-2 ratchet"
   },
   {
-    "file": "larch/implement/implement_dispatch.py",
+    "file": "larch/implement/dispatch_ship.py",
     "qualified_symbol": "ship_pre_driver_main",
     "env_name": "IMPLEMENT_TMPDIR",
     "constant": "ENV_IMPLEMENT_TMPDIR",
@@ -1134,7 +1134,7 @@
     "reason": "grandfathered bare env literal pre-G-Cfg-2 ratchet"
   },
   {
-    "file": "larch/implement/implement_dispatch.py",
+    "file": "larch/implement/dispatch_bootstrap.py",
     "qualified_symbol": "step0_bootstrap_main",
     "env_name": "IMPLEMENT_TMPDIR",
     "constant": "ENV_IMPLEMENT_TMPDIR",
@@ -1143,7 +1143,7 @@
     "reason": "grandfathered bare env literal pre-G-Cfg-2 ratchet"
   },
   {
-    "file": "larch/implement/implement_dispatch.py",
+    "file": "larch/implement/dispatch_bootstrap.py",
     "qualified_symbol": "step0_bootstrap_main",
     "env_name": "ISSUE_NUMBER",
     "constant": "ENV_ISSUE_NUMBER",
@@ -1152,7 +1152,7 @@
     "reason": "grandfathered bare env literal pre-G-Cfg-2 ratchet"
   },
   {
-    "file": "larch/implement/implement_dispatch.py",
+    "file": "larch/implement/dispatch_step2.py",
     "qualified_symbol": "step2_dispatch_main",
     "env_name": "CLAUDE_PLUGIN_ROOT",
     "constant": "ENV_CLAUDE_PLUGIN_ROOT",
@@ -1161,7 +1161,7 @@
     "reason": "grandfathered bare env literal pre-G-Cfg-2 ratchet"
   },
   {
-    "file": "larch/implement/implement_dispatch.py",
+    "file": "larch/implement/dispatch_step2.py",
     "qualified_symbol": "step2_dispatch_main",
     "env_name": "IMPLEMENT_TMPDIR",
     "constant": "ENV_IMPLEMENT_TMPDIR",
@@ -1170,7 +1170,7 @@
     "reason": "grandfathered bare env literal pre-G-Cfg-2 ratchet"
   },
   {
-    "file": "larch/implement/implement_dispatch.py",
+    "file": "larch/implement/dispatch_ship.py",
     "qualified_symbol": "step8_oos_checkpoint_main",
     "env_name": "DESIGN_TMPDIR",
     "constant": "ENV_DESIGN_TMPDIR",
@@ -1179,7 +1179,7 @@
     "reason": "grandfathered bare env literal pre-G-Cfg-2 ratchet"
   },
   {
-    "file": "larch/implement/implement_dispatch.py",
+    "file": "larch/implement/dispatch_ship.py",
     "qualified_symbol": "step8_oos_checkpoint_main",
     "env_name": "DESIGN_TMPDIR",
     "constant": "ENV_DESIGN_TMPDIR",
@@ -1188,7 +1188,7 @@
     "reason": "grandfathered bare env literal pre-G-Cfg-2 ratchet"
   },
   {
-    "file": "larch/implement/implement_dispatch.py",
+    "file": "larch/implement/dispatch_ship.py",
     "qualified_symbol": "step8_ship_main",
     "env_name": "ISSUE_NUMBER",
     "constant": "ENV_ISSUE_NUMBER",
@@ -1197,7 +1197,7 @@
     "reason": "grandfathered bare env literal pre-G-Cfg-2 ratchet"
   },
   {
-    "file": "larch/implement/implement_dispatch.py",
+    "file": "larch/implement/dispatch_ship.py",
     "qualified_symbol": "step8_ship_main",
     "env_name": "REPO",
     "constant": "ENV_REPO",
